DBLoader: bit more efficient
The header is only processed once. Still, the heavy part is the SQL
insertion, so it takes about the same time. I'm not sure of a way
to insert all together (or in batches) without making a mess or
doing it directly from the file, but that would not be in Scala...